\name{BiclustSet}
\docType{class}
\title{biclustData objects}
\description{
Class contains the biclustering result in a form: bicluster rows and bicluster columns
}
\section{Constructor}{
\describe{
\item{}{
\code{BiclustSet(...)}:
Constructs the BiclustSet object from a variety of inputs (isa2, fabia, biclust)
}
}
}
\references{
T.Khamiakova et al.(2012)
}
\author{Tatsiana Khamiakova}

\keyword{classes}