\name{superbiclust-package}
\alias{superbiclust-package}
\alias{superbiclust}
\docType{package}
\title{
generating robust biclusters form the set of biclusters

}
\description{
The package contains a number of functions for computing similarity matrix of the biclusters obtained by a variety of methods, initialization seeds or various parameter settings.
It uses biclustering output as generated by biclust or fabia. isa2 package can be used to generate the biclusters as well, however, a prior conversion is needed to a biclust object by using isa2.biclust() function. 
The matrix is used for the construction of hierarchical tree based on overall similarity, row similarity or column similarity to obtain cut-off points for the similarity metric of choice.
Various statistics are output per bicluster set: a number of a given gene(compound) or gene (compound) set has been present in any bicluster of output or per run.
After the tree is cut, the robiust or super biclusters are obtained in a form of biclust object, which can further be used for plotting of biclusters.
Biclusters are submatrices in the data which satisfy certain conditions of homogeneity. For more details on biclusters and biclustering see Madeira and Oliveira (2004).}
\details{
\tabular{ll}{
Package: \tab superbiclust\cr
Type: \tab Package\cr
Version: \tab 0.99\cr
Date: \tab 2012-08-23\cr
License: \tab GPL\cr
LazyLoad: \tab yes\cr
}
}
\author{
Tatsiana Khamiakova <tatsiana.khamiakova@uhasselt.be>
}
\references{
Madeira and Oliveira (2004) Biclustering algorithms for biological data analysis: a survey. IEEE/ACM Trans Comput Biol Bioinform. 2004 Jan-Mar;1(1):24-45.
Shi et al. (2010) A bi-ordering approach to linking gene expression	with clinical annotations in gastric cancer. BMC Bioinformatics. 11. pages 477. 
}
\concept{biclustering} 
\concept{similarity} 