\name{getStats}
\alias{getStats}
\title{
Get frequency statistic for the columns and rows membership
}
\description{
For a given Bicluster set, for each row and column in data, compute frequency of apperance within a bicluster
}
\usage{
getStats(x)
}

\arguments{
  \item{x}{
Biclust object containing bicluster results
}
}
\value{
a list of column and row frequencies
}
\author{
Tatsiana Khamiakova
}