\name{biclustSim-package}
\alias{biclustSim-package}
\alias{biclustSim}
\docType{package}
\title{
generating super biclusters form the set of overlapping outputs

}
\description{
A set of functions for generating the similarity matrix between obtained biclusters for a given dataset. The matrix is used for the construction of hierarchical tree to obtain cut-off points. The result is a number of robust biclusters with non or low overlap.}
\details{
\tabular{ll}{
Package: \tab biclustSim\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2012-08-23\cr
License: \tab GPL\cr
LazyLoad: \tab yes\cr
}
}
\author{
Tatsiana Khamiakova
Maintainer: Tatsiana Khamiakova <tatsiana.khamiakova@uhasselt.be>
}
\references{
Shi et al.
}
\keyword{biclustering, similarity}
